Enabled the editor to be built with the "release" target.
Since the game needs to ability to import resources like textures
at runtime for the convenience of the user, the release build of
open-tabletop needs to have tools enabled. However, by default,
Godot will not allow you to build a release version of its editor,
only at best a release_debug version.

This branch will allow for this, so that exported versions of the
game are as optimised as they possibly can be, and don't include
as much as they need to. This does not mean that the editor should
be used in release mode! This is purely for allowing release export
templates to be made while having tools enabled.

This fork will also include, as a submodule, the import module made
for the game for convenience.