Error handler list now uses the List class.
Before, I was getting hard crashes on macOS when the game was exiting
due to something going wrong with the pointers in the linked list.
I have replaced the raw pointer in error_macros.cpp with the List
class from list.h, which seems to have fixed the crash.

But before I merge these changes to the tabletop-3.4.4 branch, I will
first test these changes on Windows and Linux.