Updated the custom module.
It turns out that I didn't need macOS-specific fixes to the engine,
it was actually my shoddy programming that caused the crashes.
The cause was that the ErrorReporter didn't properly remove it's
error handler, so the pointer was still in the list, but the class
itself was still freed from memory, so that pointer was now corrupt.