Added upper limit to the number of errors logged to files.
Once 100,000 errors have been logged to RotatedFileLogger, it will
stop writing any future errors to the file. In testing, this results
in a maximum of about 10MiB of error text.

While this change does not fix the root cause of
drwhut/tabletop-club#131, it should help in stopping the creation
of multi-gigabyte log files in the event the engine goes into an
error loop.