Fixed set_name error when importing OBJ files.
A different method was used to fix the error in
4c69b6181e1d99831a3eefb763b14574b2e76ce8, but it had the
unintentional side effect of shifting the names of the meshes over
by one.

Instead, this method simply does not attempt to add a mesh to the
list (upon encountering an "o" command) if a vertex has not been
registered yet. Much nicer :)