Fixed focus going outside of ScrollContainers unexpectedly.
The issue was that when the engine was deciding automatically where
the focus neighbour of a control was, it did not take ScrollContainers
into account, so controls outside of the container were technically
closer in distance to the currently focused control than the controls
within the container that were not in view.

This has being fixed by checking for ScrollContainers along the way
in determining where the 'base' control is in the current scene.
If there are no valid neighbours within the container that the focus
can move to, the function proceeds as normal, and focus goes outside
of the container.

This issue has already been reported in Godot, see:
godotengine/godot#63055

The fix was based on this PR for the given issue:
godotengine/godot#63059